Update README.md
The purpose of designing this plasmid was for a different experiment to compare effectiveness of different methods to bypass vernalisation for flowering in blueberries, for their cultivation in tropical climate. Methods of choice are grafting, genetic modification, a combination of both and hormonal treatment.
The idea is, by decreasing the production of auxin (to reduce the time for early maturation) and by overexpression of genes that initiate biosynthesis of gibberellic acid (to promote floral bud development and dormancy break) a mutant can be created that flowers without a period of vernalisation.
The goal of this experiment is to determine the most effective (biologically and economically) method of growing blueberries in tropical climates. Ideally a mutant plant could be grafted on a plant of the same family (in this case, prime candidate is vaccinium leschenaultii - indian cranberry) to ensure better nutrient supply and environmental resistance.